+++ docs/changes.rst
@@ -128,6 +128,9 @@ Released: not yet
 
 * Remove use of pydicti dictionary package in favor of NocaseDict.
 
+* set pylint disable on all uses of pdb.set_trace(). This is an issue between
+  the add-on package pdbpp and lint, not pdb.  (see issue # 751)
+
 **Known issues:**
 
 * See `list of open issues`_.